refactory(repository): 전문 검색 인덱스 적
