+++ backend/src/main/resources/db/migration/V6__add_full_text_index.sql
@@ -1,0 +1,2 @@
+ALTER TABLE template ADD FULLTEXT INDEX idx_template_fulltext (title, description);
+ALTER TABLE source_code ADD FULLTEXT INDEX idx_source_code_fulltext (content, filename);
