+++ backend/src/test/java/codezap/template/service/TemplateServiceTest.java
@@ -6,7 +6,6 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
 
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -14,7 +13,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.domain.Sort.Direction;
 import org.springframework.transaction.annotation.Transactional;
@@ -30,11 +28,7 @@ import codezap.likes.domain.Likes;
 import codezap.likes.repository.LikesRepository;
 import codezap.member.domain.Member;
 import codezap.member.repository.MemberRepository;
-import codezap.tag.domain.Tag;
-import codezap.tag.repository.TagRepository;
-import codezap.tag.repository.TemplateTagRepository;
 import codezap.template.domain.Template;
-import codezap.template.domain.TemplateTag;
 import codezap.template.dto.request.CreateSourceCodeRequest;
 import codezap.template.dto.request.CreateTemplateRequest;
 import codezap.template.dto.request.UpdateTemplateRequest;
@@ -52,13 +46,8 @@ class TemplateServiceTest {
     private CategoryRepository categoryRepository;
 
     @Autowired
-    private TemplateTagRepository templateTagRepository;
-
-    @Autowired
-    private TagRepository tagRepository;
-
-    @Autowired
     private MemberRepository memberRepository;
+
 
     @Autowired
     private LikesRepository likesRepository;
@@ -165,342 +154,59 @@ class TemplateServiceTest {
     }
 
     @Nested
-    @DisplayName("검색 기능")
-    class FindAll {
-
-        private Member member1, member2;
-        private Category category1, category2;
-        private Tag tag1, tag2;
-        private Template template1, template2, template3;
+    @DisplayName("정렬 기능 테스트")
+    class SortTest {
 
         @Test
-        @DisplayName("검색 기능: 회원 ID로 템플릿 목록 조회 성공")
-        void findAllSuccessByMemberId() {
-            saveInitialData();
-            Long memberId = member1.getId();
-            String keyword = null;
-            Long categoryId = null;
-            List<Long> tagIds = null;
-            Pageable pageable = PageRequest.of(0, 10);
+        @DisplayName("좋아요 순 정렬 테스트")
+        void sortByLikesCount() {
+            saveMembers10();
+            Category category = categoryRepository.save(CategoryFixture.getFirstCategory());
+            saveTemplates5(category);
+            likeTemplate(3L, 10L);
+            likeTemplate(5L, 7L);
+            likeTemplate(2L, 5L);
+            likeTemplate(4L, 1L);
+            likeTemplate(1L, 0L);
 
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
+            List<Template> templates = sut.findAllBy(null, "", null, null,
+                            PageRequest.of(0, 10, Sort.by(Direction.DESC, "likesCount")))
+                    .getContent();
 
-            assertAll(
-                    () -> assertThat(actual.getContent()).hasSize(2),
-                    () -> assertThat(actual.getContent())
-                            .allMatch(template -> template.getMember().getId().equals(member1.getId()))
+            assertThat(templates).containsExactly(
+                    templateRepository.fetchById(3L),
+                    templateRepository.fetchById(5L),
+                    templateRepository.fetchById(2L),
+                    templateRepository.fetchById(4L),
+                    templateRepository.fetchById(1L)
             );
         }
 
-        @Test
-        @Disabled("Pageable에 대한 null 검증이 필요함")
-        @DisplayName("검색 기능 실패: Pageable을 전달하지 않은 경우")
-        void findAllFailureWithNullPageable() {
-            saveInitialData();
-            Long memberId = member1.getId();
-            String keyword = null;
-            Long categoryId = null;
-            List<Long> tagIds = null;
-            Pageable pageable = null;
-
-            assertThatThrownBy(() -> sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable))
-                    .isInstanceOf(CodeZapException.class)
-                    .hasMessage("Pageable을 필수로 작성해야 합니다.");
-        }
-
-        @Test
-        @DisplayName("검색 기능: 키워드로 템플릿 목록 조회 성공")
-        void findAllSuccessByKeyword() {
-            saveInitialData();
-            Long memberId = null;
-            String keyword = "Template";
-            Long categoryId = null;
-            List<Long> tagIds = null;
-            Pageable pageable = PageRequest.of(0, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertAll(
-                    () -> assertThat(actual.getContent())
-                            .allMatch(template ->
-                                    template.getTitle().contains(keyword) || template.getDescription()
-                                            .contains(keyword)),
-                    () -> assertThat(actual.getContent()).hasSize(3)
-            );
-        }
-
-        @Test
-        @DisplayName("검색 기능: 카테고리 ID로 템플릿 목록 조회 성공")
-        void findAllSuccessByCategoryId() {
-            saveInitialData();
-            Long memberId = null;
-            String keyword = null;
-            Long categoryId = category1.getId();
-            List<Long> tagIds = null;
-            Pageable pageable = PageRequest.of(0, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertAll(
-                    () -> assertThat(actual.getContent()).hasSize(2),
-                    () -> assertThat(actual.getContent())
-                            .allMatch(template -> template.getCategory().getId().equals(category1.getId()))
-            );
-        }
-
-        @Test
-        @DisplayName("검색 기능: 태그 ID 목록으로 템플릿 목록 조회, 모든 태그를 가진 템플릿만 조회 성공")
-        void findAllSuccessByTagIds() {
-            saveInitialData();
-            Long memberId = null;
-            String keyword = null;
-            Long categoryId = null;
-            List<Long> tagIds = List.of(tag1.getId(), tag2.getId());
-            Pageable pageable = PageRequest.of(0, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertAll(
-                    () -> assertThat(actual.getContent())
-                            .containsExactlyInAnyOrder(
-                                    templateRepository.fetchById(1L),
-                                    templateRepository.fetchById(2L)),
-                    () -> assertThat(actual.getContent()).hasSize(2)
-            );
-        }
-
-        @Test
-        @DisplayName("검색 기능: 단일 태그 ID로 템플릿 목록 조회 성공")
-        void findAllSuccessBySingleTagId() {
-            saveInitialData();
-            Long memberId = null;
-            String keyword = null;
-            Long categoryId = null;
-            List<Long> tagIds = List.of(tag2.getId());
-            Pageable pageable = PageRequest.of(0, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertAll(
-                    () -> assertThat(actual.getContent()).containsExactlyInAnyOrder(
-                            templateRepository.fetchById(1L),
-                            templateRepository.fetchById(2L),
-                            templateRepository.fetchById(3L)),
-                    () -> assertThat(actual.getContent()).hasSize(3)
-            );
-        }
-
-        @Test
-        @DisplayName("검색 기능: 회원 ID와 키워드로 템플릿 목록 조회 성공")
-        void findAllSuccessByMemberIdAndKeyword() {
-            saveInitialData();
-            Long memberId = member1.getId();
-            String keyword = "Template";
-            Long categoryId = null;
-            List<Long> tagIds = null;
-            Pageable pageable = PageRequest.of(0, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertAll(
-                    () -> assertThat(actual.getContent()).hasSize(2),
-                    () -> assertThat(actual.getContent())
-                            .allMatch(template -> template.getMember().getId().equals(member1.getId())
-                                    && (template.getTitle().contains(keyword)
-                                    || template.getDescription().contains(keyword)))
-            );
-        }
-
-        @Test
-        @DisplayName("검색 기능: 회원 ID와 카테고리 ID로 템플릿 목록 조회 성공")
-        void findAllSuccessByMemberIdAndCategoryId() {
-            saveInitialData();
-            Long memberId = member1.getId();
-            String keyword = null;
-            Long categoryId = category1.getId();
-            List<Long> tagIds = null;
-            Pageable pageable = PageRequest.of(0, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertAll(
-                    () -> assertThat(actual.getContent()).hasSize(1),
-                    () -> assertThat(actual.getContent().get(0).getMember().getId()).isEqualTo(member1.getId()),
-                    () -> assertThat(actual.getContent().get(0).getCategory().getId()).isEqualTo(category1.getId())
-            );
-        }
-
-        @Test
-        @DisplayName("검색 기능: 회원 ID와 태그 ID 목록으로 템플릿 목록 조회 성공")
-        void findAllSuccessByMemberIdAndTagIds() {
-            saveInitialData();
-            Long memberId = member1.getId();
-            String keyword = null;
-            Long categoryId = null;
-            List<Long> tagIds = List.of(tag1.getId(), tag2.getId());
-            Pageable pageable = PageRequest.of(0, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertAll(
-                    () -> assertThat(actual.getContent()).hasSize(2),
-                    () -> assertThat(actual.getContent())
-                            .containsExactlyInAnyOrder(
-                                    templateRepository.fetchById(1L),
-                                    templateRepository.fetchById(2L))
-            );
-        }
-
-        @Test
-        @DisplayName("검색 기능: 모든 검색 기준으로 템플릿 목록 조회 성공")
-        void findAllSuccessWithAllCriteria() {
-            saveInitialData();
-            Long memberId = member1.getId();
-            String keyword = "Template";
-            Long categoryId = category1.getId();
-            List<Long> tagIds = List.of(tag1.getId(), tag2.getId());
-            Pageable pageable = PageRequest.of(0, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertAll(
-                    () -> assertThat(actual.getContent()).hasSize(1),
-                    () -> assertThat(actual.getContent()).containsExactlyInAnyOrder(templateRepository.fetchById(1L))
-            );
-        }
-
-        @Test
-        @DisplayName("검색 기능: 검색 결과가 없는 경우 빈 리스트 반환 성공")
-        void findAllSuccessWithNoResults() {
-            saveInitialData();
-            Long memberId = null;
-            String keyword = "NonExistentKeyword";
-            Long categoryId = null;
-            List<Long> tagIds = null;
-            Pageable pageable = PageRequest.of(0, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertThat(actual.getContent()).isEmpty();
-        }
-
-        @Test
-        @DisplayName("검색 기능: 두 번째 페이지 조회 성공")
-        void findAllSuccessWithSecondPage() {
-            var member = memberRepository.save(MemberFixture.getFirstMember());
-            var category = categoryRepository.save(CategoryFixture.getFirstCategory());
-            for (int i = 0; i < 15; i++) {
-                templateRepository.save(TemplateFixture.get(member, category));
-            }
-            Long memberId = member.getId();
-            String keyword = null;
-            Long categoryId = null;
-            List<Long> tagIds = null;
-            Pageable pageable = PageRequest.of(1, 10);
-
-            Page<Template> actual = sut.findAllBy(memberId, keyword, categoryId, tagIds, pageable);
-
-            assertAll(
-                    () -> assertThat(actual.getContent()).hasSize(5),
-                    () -> assertThat(actual.getContent().get(0).getId()).isEqualTo(11L)
-            );
-        }
-
-        @Nested
-        @DisplayName("정렬 기능 테스트")
-        class SortTest {
-
-            @Test
-            @DisplayName("좋아요 순 정렬 테스트")
-            void sortByLikesCount() {
-                saveMembers10();
-                Category category = categoryRepository.save(CategoryFixture.getFirstCategory());
-                saveTemplates5(category);
-                likeTemplate(3L, 10L);
-                likeTemplate(5L, 7L);
-                likeTemplate(2L, 5L);
-                likeTemplate(4L, 1L);
-                likeTemplate(1L, 0L);
-
-                List<Template> templates = sut.findAllBy(null, "", null, null,
-                                PageRequest.of(0, 10, Sort.by(Direction.DESC, "likesCount")))
-                        .getContent();
-
-                assertThat(templates).containsExactly(
-                        templateRepository.fetchById(3L),
-                        templateRepository.fetchById(5L),
-                        templateRepository.fetchById(2L),
-                        templateRepository.fetchById(4L),
-                        templateRepository.fetchById(1L)
-                );
-            }
-
-            private void saveMembers10() {
-                for (int i = 0; i < 10; i++) {
-                    memberRepository.save(new Member("name" + i, "password" + 1, "salt"));
-                }
-            }
-
-            private void saveTemplates5(Category category) {
-                for (int i = 0; i < 5; i++) {
-                    templateRepository.save(new Template(
-                            memberRepository.fetchById(1L),
-                            "title" + i,
-                            "description",
-                            category
-                    ));
-                }
-            }
-
-            private void likeTemplate(long templateId, long likesCount) {
-                for (long memberId = 1L; memberId <= likesCount; memberId++) {
-                    likesRepository.save(new Likes(
-                            null,
-                            templateRepository.fetchById(templateId),
-                            memberRepository.fetchById(memberId)
-                    ));
-                }
+        private void saveMembers10() {
+            for (int i = 0; i < 10; i++) {
+                memberRepository.save(new Member("name" + i, "password" + 1, "salt"));
             }
         }
 
-        private void saveInitialData() {
-
-            saveTwoMembers();
-            saveTwoCategory();
-            saveTwoTags();
-            saveThreeTemplates();
-            saveTemplateTags();
+        private void saveTemplates5(Category category) {
+            for (int i = 0; i < 5; i++) {
+                templateRepository.save(new Template(
+                        memberRepository.fetchById(1L),
+                        "title" + i,
+                        "description",
+                        category
+                ));
+            }
         }
 
-        private void saveTwoMembers() {
-            member1 = memberRepository.save(MemberFixture.getFirstMember());
-            member2 = memberRepository.save(MemberFixture.getSecondMember());
-        }
-
-        private void saveTwoCategory() {
-            category1 = categoryRepository.save(new Category("Category 1", member1));
-            category2 = categoryRepository.save(new Category("Category 2", member1));
-        }
-
-        private void saveTwoTags() {
-            tag1 = tagRepository.save(new Tag("Tag 1"));
-            tag2 = tagRepository.save(new Tag("Tag 2"));
-        }
-
-        private void saveThreeTemplates() {
-            template1 = templateRepository.save(TemplateFixture.get(member1, category1));
-            template2 = templateRepository.save(TemplateFixture.get(member1, category2));
-            template3 = templateRepository.save(TemplateFixture.get(member2, category1));
-        }
-
-        private void saveTemplateTags() {
-            templateTagRepository.save(new TemplateTag(template1, tag1));
-            templateTagRepository.save(new TemplateTag(template1, tag2));
-
-            templateTagRepository.save(new TemplateTag(template2, tag1));
-            templateTagRepository.save(new TemplateTag(template2, tag2));
-
-            templateTagRepository.save(new TemplateTag(template3, tag2));
+        private void likeTemplate(long templateId, long likesCount) {
+            for (long memberId = 1L; memberId <= likesCount; memberId++) {
+                likesRepository.save(new Likes(
+                        null,
+                        templateRepository.fetchById(templateId),
+                        memberRepository.fetchById(memberId)
+                ));
+            }
         }
     }
